Rename wayland_compositor{c,h} to compositor.{c,h}
