+++ src/compositor/CMakeLists.txt
@@ -11,9 +11,9 @@ set(SOURCE_FILES
     monitor.c
     monitor_mode.c
     wayland/buffer.c
+    wayland/compositor.c
     wayland/seat.c
     wayland/surface.c
-    wayland/wayland_compositor.c
     wayland/wayland_pointer.c
     wayland/wayland_shell.c
     wayland/wayland_shell_surface.c
