+++ src/compositor/module.c
@@ -42,8 +42,8 @@
 #include "compositor/internal_context.h"
 #include "compositor/monitor.h"
 #include "compositor/monitor_mode.h"
+#include "compositor/wayland/compositor.h"
 #include "compositor/wayland/seat.h"
-#include "compositor/wayland/wayland_compositor.h"
 #include "compositor/wayland/wayland_shell.h"
 #include "logger/module.h"
 #include "util/cleaner.h"
